+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Release TBD
 
   * Add product image sorting.
   * Add static features/attributes.
+  * Add Supplier Reference field to products and variants for ordering.
 
 ## v1.1.0
 Release 2020-01-22